// src/types.ts
export interface AssetGroup {
  id: number;
  groupName: string;
}

export interface AssetCategory {
  id: number;
  categoryName: string;
}

export interface AssetSubCategory {
  id: number;
  subCategoryName: string;
}

export interface UOM {
  id: number;
  uomName: string;
}

export interface WorkingStatus {
  id: number;
  code: string;
}

export interface AssetType {
  id: number;
  code: string;
}

export interface ParentAsset {
  id: number;
  assetName: string;
}

export interface AssetSubGroup {
  id?: number | string;
  [key: string]: any;
}

export interface AssetState {
  selectedAssetGroup: AssetGroup | null;
  assetGroupData: AssetGroup[];
  assetSubGroupData: any[];
  selectedAssetSubGroup: AssetSubGroup;
  assetCategoryData: AssetCategory[];
  selectedAssetCategory: AssetCategory | null;
  assetSubCategoryData: AssetSubCategory[];
  selectedAssetSubCategory: AssetSubCategory | null;
  assetName: string;
  uomData: UOM[];
  selectedUom: UOM | null;
  workingStatusData: WorkingStatus[];
  selectedWorkingStatus: WorkingStatus | null;
  assetTypeData: AssetType[];
  selectedAssetType: AssetType | null;
  parentAssetData: ParentAsset[];
  selectedParentAsset: ParentAsset | null;
  quantity: string;
  description: string;
  nonDepreciated: number;
  assetImage: string | null;
  assetImageFile: File | null;
  assetImageBase64: string | null;
}

export type AssetActionType = {
  type: keyof AssetState;
  payload: any;
};